requirements.txt
fastapi
uvicorn
diffusers
torch
Pillow
python-multipart
transformers
accelerate
safetensors
scipy
numpy